WIP merged file picker fragment with torrent loader fragment
avoids caching issues, maybe fixes #407
