+++ app/app/src/main/res/layout/fragment_torrent_processing.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools">
 
     <data></data>
 
@@ -58,14 +59,39 @@
         </androidx.constraintlayout.widget.ConstraintLayout>
 
 
-        <androidx.fragment.app.FragmentContainerView
+        <androidx.constraintlayout.widget.ConstraintLayout
             android:id="@+id/loadedLayout"
-            xmlns:android="http://schemas.android.com/apk/res/android"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            android:name="com.github.livingwithhippos.unchained.torrentfilepicker.view.TorrentFilePickerFragment"
-            android:visibility="invisible"/>
+            android:visibility="invisible">
 
+            <TextView
+                android:id="@+id/tvPickerTitle"
+                android:layout_width="0dp"
+                android:layout_height="wrap_content"
+                android:layout_marginTop="10dp"
+                android:text="@string/select_files"
+                android:textAlignment="center"
+                app:layout_constraintBottom_toTopOf="@id/rvTorrentFilePicker"
+                app:layout_constraintEnd_toEndOf="parent"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintTop_toTopOf="parent" />
+
+
+            <androidx.recyclerview.widget.RecyclerView
+                android:id="@+id/rvTorrentFilePicker"
+                tools:listitem="@layout/item_list_torrent_selection_file"
+                android:layout_width="match_parent"
+                android:layout_height="0dp"
+                android:layout_gravity="center_horizontal"
+                android:layout_marginTop="15dp"
+                android:scrollbars="vertical"
+                app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
+                app:layout_constraintBottom_toBottomOf="parent"
+                app:layout_constraintEnd_toEndOf="parent"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintTop_toBottomOf="@id/tvPickerTitle" />
+        </androidx.constraintlayout.widget.ConstraintLayout>
 
         <com.google.android.material.floatingactionbutton.FloatingActionButton
             android:id="@+id/fabDownload"
